I've added a plot for the force we're trying to simulate with the *current* settings, but of course these are just parameters
I will now try to speed up the calculations by the simple method of replacing for i, for j by for i, for j < i and calculating the
symmetric pair interactions at the same time.

Afterwards I will make sure the order/order' are actually plotted because that part is not currently working (and to be fair, I haven't looked
at it much beyond introducing the subplots).